✨[Feat] main -> profile/[userId] , intro -> main 페이지 이동 및 profile api 연동 #282
