+++ src/app/admin/(block)/image/components/image-box.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Image from "next/image";
 import Link from "next/link";
 
-import ErrorBoundary from "@app/intro/components/error-boundary";
+import ErrorBoundary from "@components/error-boundary";
 
 interface Props {
   // handeInputImageClick: () => void;
