+++ src/app/profile/[userId]/page.tsx
@@ -1,0 +1,11 @@
+import Preview from "@app/admin/components/preview/components/preview";
+
+function Page() {
+  return (
+    <div className={"px-14"}>
+      <Preview />
+    </div>
+  );
+}
+
+export default Page;
